feat : app.tsx view로 분리
